requirements-sensor.txt
bmp280
python-dotenv
requests
